Merge commit 'ce15381fb29f3c8e72cc86e1e69f906e9cfd865d' into feature/websocketnavigation

* commit 'ce15381fb29f3c8e72cc86e1e69f906e9cfd865d': (21 commits)
  Fix browser mask bug introduced when flipping y axis
  Fix crash related to deinitialization of screenspacebrowser
  Feature/multiple endpoints (#938)
  Add missing asset file
  Fix scaling issue with screenspace renderables (#939)
  Compile fix if instrumentation is disabled
  Add basic instrumentation options to the renderengine and globebrowsing module Instrumentation is disabled on default, but can be enabled in CMake
  Update Apollo assets
  Remove torrent-based synchronization
  Cleanup of Wormhole application Updated Ghoul to fix commandline argument extraction bug
  Make constructor of Time explicit
  Feature/timeline translation (#931)
  Use correct identifier for Vesta
  Fix copy and paste error
  minVR updates (#908)
  Revive screen space browser and add example (#932)
  Feature/navigation state (#930)
  Fixes for the Gaia data sync
  Add support for arrays in luascript topic (#913)
  Feature/flying fixes (#924)
  ...

# Conflicts:
#	data/assets/default.scene
#	include/openspace/interaction/navigationhandler.h
#	include/openspace/interaction/orbitalnavigator.h
#	src/interaction/orbitalnavigator.cpp

+++ modules/base/dashboard/dashboarditemframerate.cpp
@@ -289,7 +289,7 @@ void DashboardItemFramerate::render(glm::vec2& penPosition) {
     );
 
     int nLines = output.empty() ? 0 :
-        (std::count(output.begin(), output.end(), '\n') + 1);
+        static_cast<int>((std::count(output.begin(), output.end(), '\n') + 1));
 
     penPosition.y -= _font->height() * static_cast<float>(nLines);
